Version Packages (#391)
This PR was opened by the [Changesets release](https://github.com/changesets/action) GitHub action. When you're ready to do a release, you can merge this and the packages will be published to npm automatically. If you're not ready to do a release yet, that's fine, whenever you add more changesets to main, this PR will be updated.


# Releases
## ai@2.1.27

### Patch Changes

-   3d29799: React/Svelte/Vue: keep isLoading in sync between hooks with the same ID.

    React: don't throw error when submitting